refactor: move Importer class functionality into importer.main file
