+++ scopes/scope/importer/import.cmd.ts
@@ -17,7 +17,7 @@ import ConsumerComponent from '@teambit/legacy/dist/consumer/component/consumer-
 import GeneralError from '@teambit/legacy/dist/error/general-error';
 import { immutableUnshift } from '@teambit/legacy/dist/utils';
 import { formatPlainComponentItem } from '@teambit/legacy/dist/cli/chalk-box';
-import { Importer } from './importer';
+import { ImporterMain } from './importer.main.runtime';
 
 export default class ImportCmd implements Command {
   name = 'import [component-patterns...]';
@@ -73,7 +73,7 @@ export default class ImportCmd implements Command {
   remoteOp = true;
   _packageManagerArgs: string[]; // gets populated by yargs-adapter.handler().
 
-  constructor(private importer: Importer, private docsDomain: string) {
+  constructor(private importer: ImporterMain, private docsDomain: string) {
     this.extendedDescription = `https://${docsDomain}/components/importing-components
 ${WILDCARD_HELP('import')}`;
   }
