+++ scopes/scope/importer/index.ts
@@ -1,6 +1,5 @@
 import { ImporterAspect } from './importer.aspect';
 
 export type { ImporterMain } from './importer.main.runtime';
-export { Importer } from './importer';
 export default ImporterAspect;
 export { ImporterAspect };
